requirements.txt
voyageai
numpy
requests
anthropic
python-telegram-bot
base58

